backend/pyproject.toml
[project]
name = "tugan-vpn"
version = "0.1.0"
description = ""
authors = [
    {name = "miderror",email = "miderror.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "yookassa (>=3.7.1,<4.0.0)",
    "py3xui (>=0.5.2,<0.6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "telegram-webapp-auth (>=3.0.6,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-jsonform (>=2.23.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
